+++ Hinotori/Administrador/src/main/java/Administrador/model/dao/CidadeDao.java
@@ -8,8 +8,8 @@ public interface CidadeDao {
 
 	void insert(Cidade obj);
 	void update(Cidade obj);
-	void delete(Integer id);
-	Cidade find(Integer id);
+	void delete(Long id);
+	Cidade find(Long id);
 	List<Cidade> findAll();
 	
 }
